Support deleted records
Add support for deleted records. Those records should have a state that
signifies that the record is deleted but the record itself should be
still indexed in Solr. In this way it is possible to comply to the
transient and persistent delete levels defined in OAI-PMH.
The deleted records should be filtered out of the normal search results
bij for example base filters. This base filter will be removed when
searching for harvesting records, if it has the following format:
-solrfield:deletevalue